Initial file structure refactoring
