+++ hook-emulator/lib/account-manager.js
